backend/requirements/base.txt
Django==3.1.1
django-environ==0.4.5
psycopg2==2.8.5
djangorestframework==3.11.1
django-cors-headers==3.5.0
Pillow==7.2.0
